fixed some primary keys data types + made a reservableRepositoryByName

+++ server/src/main/java/nl/tudelft/oopp/server/repositories/FoodCourtRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
  * interface to implement crud operations for food court.
  */
 @Repository
-public interface FoodCourtRepository extends JpaRepository<FoodCourt,Long> {
+public interface FoodCourtRepository extends JpaRepository<FoodCourt,Integer> {
 }
